Add Dockerfile, compose.yaml, and an unbound example config
This adds a Dockerfile for unbound exporter, which we can publish in the
future.  An example docker compose.yml is included to demonstrate and
test using it with unbound, along with a sample configuration file for
unbound showing how to set up the remote-control.

The unbound example config file is based on the one inside the
mvance/docker image that's used here.