Add version collector
Adds a new unbound_exporter_build_info metric.
Most of our other Go packages include this.

It's useful for monitoring deployed versions of code via Prometheus.

In the dockerfile, copy in all files so that the toolchain knows what
git commit it is being built from (instead of just *.go)