Fix panic on mem.quic with a refactor for testability (#105)
This refactors the exporter to keep an internal set of compiled metrics,
instead of having them in a global variable. That means I can write tests
using what was the input to newUnboundMetric.

Since it's not a global anymore, it has to be passed into collectFromSocket
and collectFromReader. These aren't a meaningful part of the API, so make them
private before changing their signatures. I've made collectFromSocket a member
of UnboundExporter because I think in a future change we can also hold onto the
socket longer.

This adds a new test that motivated the change, which verifies that the number
of labels in each metric description matches the number of subexpressions in
the regex, and a fix to the underlying bug.

Fixes #104